Move rmq properties to a separate class
