+++ src/test/java/com/ericsson/ei/flowtests/TrafficGeneratedTest.java
@@ -55,6 +55,7 @@ import com.ericsson.ei.erqueryservice.ERQueryService;
 import com.ericsson.ei.erqueryservice.SearchOption;
 import com.ericsson.ei.handlers.ObjectHandler;
 import com.ericsson.ei.handlers.RMQHandler;
+import com.ericsson.ei.handlers.RMQProperties;
 import com.ericsson.ei.handlers.UpStreamEventsHandler;
 import com.ericsson.ei.test.utils.TestConfigs;
 import com.ericsson.ei.utils.TestContextInitializer;
@@ -123,7 +124,8 @@ public class TrafficGeneratedTest extends FlowTestBase {
             List<String> events = getPreparedEventsToSend(eventNames);
             int eventsCount = eventNames.size() * EVENT_PACKAGES;
 
-            String queueName = rmqHandler.getQueueName();
+            RMQProperties rmqProperties = rmqHandler.getRmqProperties();
+            String queueName = rmqProperties.getQueueName();
             String exchange = "ei-poc-4";
             TestConfigs.createExchange(exchange, queueName);
             Channel channel = TestConfigs.getConnection().createChannel();
